# brag-output/work/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: compliance-engine API
  version: 1.0.0
  description: >-
    Versioned REST API of the compliance engine. Every authenticated request is
    scoped to the tenant workspace bound to its bearer token. Outputs are
    operational readiness aids, not legal advice and not certification.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/members:
    get:
      tags: [access]
      operationId: listMembers
      x-permission: members.manage
      summary: Members of the workspace and their roles
      responses:
        "200":
          description: Members ordered by email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  members: {type: array, items: {$ref: "#/components/schemas/Member"}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
  /api/v1/members/{email}:
    put:
      tags: [access]
      operationId: putMember
      parameters:
        - {name: email, in: path, required: true, schema: {type: string, format: email}}
      x-permission: members.manage
      summary: Grant roles to a user in the workspace (creates a user without a password when needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              additionalProperties: false
              properties:
                roles: {type: array, minItems: 1, items: {$ref: "#/components/schemas/Role"}}
      responses:
        "200":
          description: The membership.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Member"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
    delete:
      tags: [access]
      operationId: deleteMember
      x-permission: members.manage
      summary: Remove a member and revoke their tokens in the workspace
      parameters:
        - {name: email, in: path, required: true, schema: {type: string, format: email}}
      responses:
        "204": {description: Removed.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
  /api/v1/tokens:
    get:
      tags: [access]
      operationId: listTokens
      x-permission: tokens.own
      summary: API tokens of the workspace (admins) or the caller's own tokens
      responses:
        "200":
          description: Tokens without their values or hashes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens: {type: array, items: {$ref: "#/components/schemas/Token"}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
    post:
      tags: [access]
      operationId: postToken
      x-permission: tokens.own
      summary: Create an API token; its value is returned once
      description: >-
        A user token acts as the caller and may carry only roles the caller holds.
        Service tokens (service=true) have no user, may hold only owner and auditor,
        and require the admin role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, roles]
              additionalProperties: false
              properties:
                name: {type: string, minLength: 1, maxLength: 100}
                roles: {type: array, minItems: 1, items: {$ref: "#/components/schemas/Role"}}
                expires_at: {type: string, format: date-time}
                service: {type: boolean}
      responses:
        "201":
          description: The token and its value. The value is never shown again.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: {$ref: "#/components/schemas/Token"}
                  value: {type: string}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
  /api/v1/tokens/{id}:
    delete:
      tags: [access]
      operationId: deleteToken
      x-permission: tokens.own
      summary: Revoke a token (admins any token of the workspace, others their own)
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "204": {description: Revoked.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
  /api/v1/audit:
    get:
      tags: [access]
      operationId: getAudit
      x-permission: audit.read
      summary: The workspace's append-only audit log, optionally with chain verification
      parameters:
        - {name: after_seq, in: query, schema: {type: integer, minimum: 0}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000, default: 100}}
        - {name: verify, in: query, schema: {type: boolean}}
      responses:
        "200":
          description: Events in ascending order.
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: {type: array, items: {$ref: "#/components/schemas/AuditEvent"}}
                  next_after_seq: {type: integer}
                  verification: {$ref: "#/components/schemas/AuditVerification"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        API token bound to one tenant workspace and a set of roles: a static token
        configured in COMPLIANCE_TOKENS ('compliance-engine token generate'), or a
        stored token created through POST /api/v1/tokens or 'compliance-engine token create'.
        Every operation declares the permission it requires in x-permission.
...
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unauthorized:
      description: Missing or invalid bearer token.
      headers:
        WWW-Authenticate: {schema: {type: string}}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Forbidden:
      description: >-
        The caller's roles lack the operation's permission (code forbidden or
        role_not_held), or the feature is not entitled (code feature_not_entitled).
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    RateLimited:
      description: Too many requests from this caller, or too many failed authentications from this address.
      headers:
        Retry-After: {schema: {type: integer}, description: Seconds to wait.}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Finding:
      type: object
      properties:
        severity: {type: string, enum: [error, warning]}
        code: {type: string}
        message: {type: string}
        template: {type: string}
        row: {type: string}
        field: {type: string}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [unauthorized, invalid_json, invalid_parameter, missing_parameter, feature_not_entitled, not_found, conflict,
                     already_rolled_back, nothing_to_roll_back, payload_too_large, too_many_records, invalid_batch,
                     unsupported_schema_version, unknown_adapter, mode_not_supported, invalid_manifest, name_mismatch, invalid_file,
                     forbidden, role_not_held, last_admin, invalid_request, already_exists, rate_limited, identity_not_configured,
                     invalid_transition, stale_evaluation, self_approval_forbidden, evidence_required, not_control_owner,
                     cannot_verify_here, already_revoked, no_checksum, managed_storage_not_configured, not_managed, export_incomplete,
                     recommendation_required, submitter_approval_forbidden, already_approved, workspace_suspended,
                     workspace_limit_reached, evidence_retention, internal]
            message: {type: string}
            details:
              type: array
              items: {$ref: "#/components/schemas/FieldError"}
            feature: {type: string}
            reason: {type: string, enum: [not_licensed, expired, addon_disabled]}
            permission: {type: string, description: The permission a forbidden request lacked.}
            findings:
              type: array
              description: The blocking findings of a refused incomplete export (export_incomplete).
              items: {$ref: "#/components/schemas/Finding"}
    Role:
      type: string
      enum: [auditor, owner, reviewer, approver, admin]
...
    Member:
      type: object
      properties:
        scope: {type: object}
        user_id: {type: string}
        email: {type: string}
        roles: {type: array, items: {$ref: "#/components/schemas/Role"}}
        updated_at: {type: string, format: date-time}
    Token:
      type: object
      properties:
        id: {type: string}
        scope: {type: object}
        name: {type: string}
        user_id: {type: string, description: Absent for service tokens.}
        roles: {type: array, items: {$ref: "#/components/schemas/Role"}}
        created_by: {type: string}
        created_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        revoked_at: {type: string, format: date-time}
        last_used_at: {type: string, format: date-time}
    AuditEvent:
      type: object
      properties:
        scope: {type: object}
        seq: {type: integer}
        at: {type: string, format: date-time}
        actor: {type: string}
        actor_kind: {type: string}
        action: {type: string}
        target_type: {type: string}
        target_id: {type: string}
        details: {type: object}
        prev_hash: {type: string}
        hash: {type: string, description: "SHA-256 over prev_hash and the event; see docs/access.md"}
    AuditVerification:
      type: object
      properties:
        ok: {type: boolean}
        checked: {type: integer}
        broken_at: {type: integer}
        reason: {type: string, enum: [sequence_gap, prev_hash_mismatch, hash_mismatch]}
    FieldError:
      type: object
      properties:
        entity: {type: string}
        index: {type: integer}
        source_record_ref: {type: string}
        field: {type: string}
        code: {type: string}
        message: {type: string}
...